transaction-api/requirements.txt
fastapi
uvicorn[standard]
kafka-python
pydantic
python-dotenv
databases[postgresql]
asyncpg
